create dynamicLayoutStore to separate dynamic layout from LayoutStore
